Implement support for a deadman's switch.
Deadman's switches are alerts that are always firing. This way, we can check
that the alerting stack is functional. We never want to receive an alert for
a deadman's switch, but we want to receive an alert when it is missing.

This commit adds the -deadman-switch-name option. An alert with that alertname
is removed from incoming posts. If the alert is the only one firing, the
overall status is overwritten to "resolved". If the alert is not firing, a new
alert is added and the overall status is set to "firing".